person-agent/pyproject.toml
[project]
name = "person-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai>=0.8.0",
    "langchain>=0.2.0",
    "langchain-groq>=0.1.0",
    "langchain-mcp-adapters>=0.1.0",
    "langgraph>=0.1.0",
    "playwright>=1.40.0",
    "fastmcp>=0.1.0",
    "streamlit>=1.30.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "langchain-openai>=0.1.0",
]
